+++ estampes/parser/gaussian/glog/logkeys.py
@@ -3,6 +3,9 @@
 To limit redundancy, common or complex patters are stored as variables.
 """
 
+KEY_UINT = r'\s*\d+'
+KEY_FP = r'\s*-?\d+\.\d+'
+
 # RR_OMEGA = r'^\s+-> (?P<val>Omega =\s*\d+\.\d+ cm.-1.*?)\s*$'
 # RR_OMEGA_LINE = r'^\s+-> (?P<val>Omega =\s*\d+\.\d+ cm.-1' \
 #     + r'(?:, Gamma =\s*\d+\.\d+ cm.-1))?\s*$'